pyproject.toml
[project]
name = "ccc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.1.1",
    "slack-sdk>=3.33.6",
    "requests>=2.32.3",
    "python-slugify>=8.0.0",
    "pytest>=8.4.2",
    "pygithub>=2.8.1",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "asyncio: marks tests as async (pytest-asyncio)",
]
asyncio_mode = "auto"
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
]
